It finally works... I can't believe it finally works. Of all the biggest pains in the arse, this would be the one that takes the cake. All the 3D tests pass.
The MEG6 scheme now uses _only_ the domain data, so there doesn't need to be coordinate data in the halo regions. This bodes well for non-functionally defined coordinate grids. Down the road coordinates will be provided as arrays, rather than as functions, so adaptation could be used in the future.

Next comes cleanup!